Home

ModernRobotics.jl

ModernRobotics.Adjoint

ModernRobotics.AxisAng3

ModernRobotics.AxisAng6

ModernRobotics.ComputedTorque

ModernRobotics.CubicTimeScaling

ModernRobotics.DistanceToSE3

ModernRobotics.DistanceToSO3

ModernRobotics.EndEffectorForces

ModernRobotics.EulerStep

ModernRobotics.FKinBody

ModernRobotics.FKinSpace

ModernRobotics.ForwardDynamics

ModernRobotics.ForwardDynamicsTrajectory

ModernRobotics.GravityForces

ModernRobotics.IKinBody

ModernRobotics.IKinSpace

ModernRobotics.InverseDynamics

ModernRobotics.InverseDynamicsTrajectory

ModernRobotics.JacobianBody

ModernRobotics.JacobianSpace

ModernRobotics.JointTrajectory

ModernRobotics.MassMatrix

ModernRobotics.MatrixExp3

ModernRobotics.MatrixExp6

ModernRobotics.MatrixLog3

ModernRobotics.MatrixLog6

ModernRobotics.NearZero

ModernRobotics.Normalize

ModernRobotics.ProjectToSE3

ModernRobotics.ProjectToSO3

ModernRobotics.QuinticTimeScaling

ModernRobotics.RotInv

ModernRobotics.RpToTrans

ModernRobotics.ScrewToAxis

ModernRobotics.SimulateControl

ModernRobotics.TestIfSE3

ModernRobotics.TestIfSO3

ModernRobotics.TransInv

ModernRobotics.TransToRp

ModernRobotics.VecTose3

ModernRobotics.VecToso3

ModernRobotics.VelQuadraticForces

ModernRobotics.ad

ModernRobotics.se3ToVec

ModernRobotics.so3ToVec

ModernRobotics.Adjoint — Method.

Adjoint(T)

Computes the adjoint representation of a homogeneous transformation matrix.

Examples

```jldoctest julia> Adjoint([1 0 0 0; 0 0 -1 0; 0 1 0 3; 0 0 0 1]) 6×6 Array{Float64,2}: 1.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 -1.0 0.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 0.0 0.0 0.0 3.0 1.0 0.0 0.0 3.0 0.0 0.0 0.0 0.0 -1.0 0.0 0.0 0.0 0.0 1.0 0.0

source

ModernRobotics.AxisAng3 — Method.

AxisAng3(expc3)

Converts a 3-vector of exponential coordinates for rotation into axis-angle form.

Examples

```jldoctest julia> AxisAng3([1 2 3]) ([0.267261 0.534522 0.801784], 3.7416573867739413)

source

ModernRobotics.AxisAng6 — Method.

AxisAng6(expc6)

Converts a 6-vector of exponential coordinates into screw axis-angle form.

Examples

```jldoctest julia> AxisAng6([1, 0, 0, 1, 2, 3]) ([1.0, 0.0, 0.0, 1.0, 2.0, 3.0], 1.0)

source

ModernRobotics.ComputedTorque — Method.

ComputedTorque(thetalist, dthetalist, eint, g, Mlist, Glist, Slist, thetalistd, dthetalistd, ddthetalistd, Kp, Ki, Kd)

Computes the joint control torques at a particular time instant.

source

ModernRobotics.CubicTimeScaling — Method.

CubicTimeScaling(Tf, t)

Computes s(t) for a cubic time scaling.

Examples

```jldoctest julia> CubicTimeScaling(2, 0.6) 0.21600000000000003

source

ModernRobotics.DistanceToSE3 — Method.

DistanceToSE3(mat)

Returns the Frobenius norm to describe the distance of mat from the SE(3) manifold.

Examples

```jldoctest julia> DistanceToSE3([1.0 0.0 0.0 1.2; 0.0 0.1 -0.95 1.5; 0.0 1.0 0.1 -0.9; 0.0 0.0 0.1 0.98]) 0.13493053768513638

source

ModernRobotics.DistanceToSO3 — Method.

DistanceToSO3(mat)

Returns the Frobenius norm to describe the distance of mat from the SO(3) manifold.

Examples

```jldoctest julia> DistanceToSO3([1.0 0.0 0.0 ; 0.0 0.1 -0.95; 0.0 1.0 0.1]) 0.08835298523536149

source

ModernRobotics.EndEffectorForces — Method.

EndEffectorForces(thetalist, Ftip, Mlist, Glist, Slist)

Computes the joint forces/torques an open chain robot requires only to create the end-effector force Ftip.

Examples

```jldoctest julia> EndEffectorForces(thetalist, Ftip, Mlist, Glist, Slist) 3-element Array{Float64,1}: 1.4095460782639782 1.8577149723180628 1.392409

source

ModernRobotics.EulerStep — Method.

EulerStep(thetalist, dthetalist, ddthetalist, dt)

Compute the joint angles and velocities at the next timestep using first order Euler integration.

Examples

```jldoctest julia> EulerStep(thetalist, dthetalist, ddthetalist, dt) ([0.11, 0.12, 0.13], [0.3, 0.35, 0.4])

source

ModernRobotics.FKinBody — Method.

FKinBody(M, Blist, thetalist)

Computes forward kinematics in the body frame for an open chain robot.

Examples

```jldoctest julia> FKinBody(M, Blist, thetalist) 4×4 Array{Float64,2}: -1.14424e-17 1.0 0.0 -5.0 1.0 1.14424e-17 0.0 4.0 0.0 0.0 -1.0 1.68584 0.0 0.0 0.0 1.0

source

ModernRobotics.FKinSpace — Method.

FKinSpace(M, Slist, thetalist)

Computes forward kinematics in the space frame for an open chain robot.

Examples

```jldoctest julia> FKinSpace(M, Slist, thetalist) 4×4 Array{Float64,2}: -1.14424e-17 1.0 0.0 -5.0 1.0 1.14424e-17 0.0 4.0 0.0 0.0 -1.0 1.68584 0.0 0.0 0.0 1.0

source

ModernRobotics.ForwardDynamics — Method.

ForwardDynamics(thetalist, dthetalist, taulist, g, Ftip, Mlist, Glist, Slist)

Computes forward dynamics in the space frame for an open chain robot.

Examples

```jldoctest julia> ForwardDynamics(thetalist, dthetalist, taulist, g, Ftip, Mlist, Glist, Slist) 3-element Array{Float64,1}: -0.9739290670855626 25.584667840340558 -32.91499212478149

source

ModernRobotics.ForwardDynamicsTrajectory — Method.

ForwardDynamicsTrajectory(thetalist, dthetalist, taumat, g, Ftipmat, Mlist, Glist, Slist, dt, intRes)

Simulates the motion of a serial chain given an open-loop history of joint forces/torques.

source

ModernRobotics.GravityForces — Method.

GravityForces(thetalist, g, Mlist, Glist, Slist)

Computes the joint forces/torques an open chain robot requires to overcome gravity at its configuration.

Examples

```jldoctest julia> GravityForces(thetalist, g, Mlist, Glist, Slist) 3-element Array{Float64,1}: 28.40331261821983 -37.64094817177068 -5.4415891999683605

source

ModernRobotics.IKinBody — Method.

IKinBody(Blist, M, T, thetalist0, eomg, ev)

Computes inverse kinematics in the body frame for an open chain robot.

Examples

```jldoctest julia> IKinBody(Blist, M, T, thetalist0, eomg, ev) ([1.57074; 2.99967; 3.14154], true)

source

ModernRobotics.IKinSpace — Method.

IKinSpace(Slist, M, T, thetalist0, eomg, ev)

Computes inverse kinematics in the space frame for an open chain robot.

Examples

```jldoctest julia> IKinSpace(Slist, M, T, thetalist0, eomg, ev) ([1.57074; 2.99966; 3.14153], true)

source

ModernRobotics.InverseDynamics — Method.

InverseDynamics(thetalist, dthetalist, ddthetalist, g, Ftip, Mlist, Glist, Slist)

Computes inverse dynamics in the space frame for an open chain robot.

Examples

```jldoctest julia> InverseDynamics(thetalist, dthetalist, ddthetalist, g, Ftip, Mlist, Glist, Slist) 3-element Array{Float64,1}: 74.69616155287451 -33.06766015851458 -3.230573137901424

source

ModernRobotics.InverseDynamicsTrajectory — Method.

InverseDynamicsTrajectory(thetamat, dthetamat, ddthetamat, g, Ftipmat, Mlist, Glist, Slist)

Calculates the joint forces/torques required to move the serial chain along the given trajectory using inverse dynamics.

source

ModernRobotics.JacobianBody — Method.

JacobianBody(Blist, thetalist)

Computes the body Jacobian for an open chain robot.

Examples

```jldoctest julia> JacobianBody(Blist, thetalist) 6×4 Array{Float64,2}: -0.0452841 0.995004 0.0 1.0 0.743593 0.0930486 0.362358 0.0 -0.667097 0.0361754 -0.932039 0.0 2.32586 1.66809 0.564108 0.2 -1.44321 2.94561 1.43307 0.3 -2.0664 1.82882 -1.58869 0.4

source

ModernRobotics.JacobianSpace — Method.

JacobianSpace(Slist, thetalist)

Computes the space Jacobian for an open chain robot.

Examples

```jldoctest julia> JacobianSpace(Slist, thetalist) 6×4 Array{Float64,2}: 0.0 0.980067 -0.0901156 0.957494 0.0 0.198669 0.444554 0.284876 1.0 0.0 0.891207 -0.0452841 0.0 1.95219 -2.21635 -0.511615 0.2 0.436541 -2.43713 2.77536 0.2 2.96027 3.23573 2.22512

source

ModernRobotics.JointTrajectory — Method.

JointTrajectory(thetastart, thetaend, Tf, N, method)

Computes a straight-line trajectory in joint space.

source

ModernRobotics.MassMatrix — Method.

MassMatrix(thetalist, Mlist, Glist, Slist)

Computes the mass matrix of an open chain robot based on the given configuration.

Examples

```jldoctest julia> MassMatrix(thetalist, Mlist, Glist, Slist) 3×3 Array{Float64,2}: 22.5433 -0.307147 -0.00718426 -0.307147 1.96851 0.432157 -0.00718426 0.432157 0.191631

source

ModernRobotics.MatrixExp3 — Method.

MatrixExp3(so3mat)

Computes the matrix exponential of a matrix in so(3).

Examples

```jldoctest julia> MatrixExp3([0 -3 2; 3 0 -1; -2 1 0]) 3×3 Array{Float64,2}: -0.694921 0.713521 0.0892929 -0.192007 -0.303785 0.933192 0.692978 0.63135 0.348107

source

ModernRobotics.MatrixExp6 — Method.

MatrixExp6(se3mat)

Computes the matrix exponential of an se3 representation of exponential coordinates.

Examples

```jldoctest julia> MatrixExp6([0 0 0 0; 0 0 -1.57079632 2.35619449; 0 1.57079632 0 2.35619449; 0 0 0 0]) 4×4 Array{Float64,2}: 1.0 0.0 0.0 0.0 0.0 6.7949e-9 -1.0 1.01923e-8 0.0 1.0 6.7949e-9 3.0 0.0 0.0 0.0 1.0

source

ModernRobotics.MatrixLog3 — Method.

MatrixLog3(R)

Computes the matrix logarithm of a rotation matrix.

Examples

```jldoctest julia> MatrixLog3([0 0 1; 1 0 0; 0 1 0]) 3×3 Array{Float64,2}: 0.0 -1.2092 1.2092 1.2092 0.0 -1.2092 -1.2092 1.2092 0.0

source

ModernRobotics.MatrixLog6 — Method.

MatrixLog6(T)

Computes the matrix logarithm of a homogeneous transformation matrix.

Examples

```jldoctest julia> MatrixLog6([1 0 0 0; 0 0 -1 0; 0 1 0 3; 0 0 0 1]) 4×4 Array{Float64,2}: 0.0 0.0 0.0 0.0 0.0 0.0 -1.5708 2.35619 0.0 1.5708 0.0 2.35619 0.0 0.0 0.0 0.0

source

ModernRobotics.NearZero — Method.

NearZero(z)

Determines whether a scalar is small enough to be treated as zero.

Examples

julia> NearZero(-1e-7)
true

source

ModernRobotics.Normalize — Method.

Normalize(V)

Normalizes a vector.

Examples

```jldoctest julia> Normalize([1 2 3]) 1×3 Array{Float64,2}: 0.267261 0.534522 0.801784

source

ModernRobotics.ProjectToSE3 — Method.

ProjectToSE3(mat)

Returns a projection of mat into SE(3).

Examples

```jldoctest julia> ProjectToSE3([0.675 0.150 0.720 1.2; 0.370 0.771 -0.511 5.4; -0.630 0.619 0.472 3.6; 0.003 0.002 0.010 0.9]) 4×4 Array{Float64,2}: 0.679011 0.148945 0.718859 1.2 0.373207 0.773196 -0.512723 5.4 -0.632187 0.616428 0.469421 3.6 0.0 0.0 0.0 1.0

source

ModernRobotics.ProjectToSO3 — Method.

ProjectToSO3(mat)

Returns a projection of mat into SO(3).

Examples

```jldoctest julia> ProjectToSO3([0.675 0.150 0.720; 0.370 0.771 -0.511; -0.630 0.619 0.472]) 3×3 Array{Float64,2}: 0.679011 0.148945 0.718859 0.373207 0.773196 -0.512723 -0.632187 0.616428 0.469421

source

ModernRobotics.QuinticTimeScaling — Method.

QuinticTimeScaling(Tf, t)

Computes s(t) for a quintic time scaling.

Examples

```jldoctest julia> QuinticTimeScaling(2, 0.6) 0.16308

source

ModernRobotics.RotInv — Method.

RotInv(R)

Inverts a rotation matrix.

Examples

```jldoctest julia> RotInv([0 0 1; 1 0 0; 0 1 0]) 3×3 LinearAlgebra.Adjoint{Int64,Array{Int64,2}}: 0 1 0 0 0 1 1 0 0

source

ModernRobotics.RpToTrans — Method.

RpToTrans(R, p)

Converts a rotation matrix and a position vector into homogeneous transformation matrix.

Examples

```jldoctest julia> RpToTrans([1 0 0; 0 0 -1; 0 1 0], [1, 2, 5]) 4×4 Array{Int64,2}: 1 0 0 1 0 0 -1 2 0 1 0 5 0 0 0 1

source

ModernRobotics.ScrewToAxis — Method.

ScrewToAxis(q, s, h)

Takes a parametric description of a screw axis and converts it to a normalized screw axis.

Examples

```jldoctest julia> ScrewToAxis([3; 0; 0], [0; 0; 1], 2) 6-element Array{Int64,1}: 0 0 1 0 -3 2

source

ModernRobotics.SimulateControl — Method.

SimulateControl(thetalist, dthetalist, g, Ftipmat, Mlist, Glist,
                Slist, thetamatd, dthetamatd, ddthetamatd, gtilde,
                Mtildelist, Gtildelist, Kp, Ki, Kd, dt, intRes)

Simulates the computed torque controller over a given desired trajectory.

source

ModernRobotics.TestIfSE3 — Method.

TestIfSE3(mat)

Returns true if mat is close to or on the manifold SE(3).

Examples

```jldoctest julia> TestIfSE3([1.0 0.0 0.0 1.2; 0.0 0.1 -0.95 1.5; 0.0 1.0 0.1 -0.9; 0.0 0.0 0.1 0.98]) false

source

ModernRobotics.TestIfSO3 — Method.

TestIfSO3(mat)

Returns true if mat is close to or on the manifold SO(3).

Examples

```jldoctest julia> TestIfSO3([1.0 0.0 0.0; 0.0 0.1 -0.95; 0.0 1.0 0.1]) false

source

ModernRobotics.TransInv — Method.

TransInv(T)

Inverts a homogeneous transformation matrix.

Examples

```jldoctest julia> TransInv([1 0 0 0; 0 0 -1 0; 0 1 0 3; 0 0 0 1]) 4×4 Array{Int64,2}: 1 0 0 0 0 0 1 -3 0 -1 0 0 0 0 0 1

source

ModernRobotics.TransToRp — Method.

TransToRp(T)

Converts a homogeneous transformation matrix into a rotation matrix and position vector.

Examples

```jldoctest julia> TransToRp([1 0 0 0; 0 0 -1 0; 0 1 0 3; 0 0 0 1]) ([1 0 0; 0 0 -1; 0 1 0], [0, 0, 3])

source

ModernRobotics.VecTose3 — Method.

VecTose3(V)

Converts a spatial velocity vector into a 4x4 matrix in se3.

Examples

```jldoctest julia> VecTose3([1 2 3 4 5 6]) 4×4 Array{Float64,2}: 0.0 -3.0 2.0 4.0 3.0 0.0 -1.0 5.0 -2.0 1.0 0.0 6.0 0.0 0.0 0.0 0.0

source

ModernRobotics.VecToso3 — Method.

VecToso3(omg)

Converts a 3-vector to an so(3) representation.

Examples

```jldoctest julia> VecToso3([1 2 3]) 3×3 Array{Int64,2}: 0 -3 2 3 0 -1 -2 1 0

source

ModernRobotics.VelQuadraticForces — Method.

VelQuadraticForces(thetalist, dthetalist, Mlist, Glist, Slist)

Computes the Coriolis and centripetal terms in the inverse dynamics of an open chain robot.

Examples

```jldoctest julia> VelQuadraticForces(thetalist, dthetalist, Mlist, Glist, Slist) 3-element Array{Float64,1}: 0.26453118054501235 -0.0550515682891655 -0.006891320068248911

source

ModernRobotics.ad — Method.

ad(V)

Calculate the 6x6 matrix [adV] of the given 6-vector.

Examples

```jldoctest julia> ad([1, 2, 3, 4, 5, 6]) 6×6 Array{Float64,2}: 0.0 -3.0 2.0 0.0 0.0 0.0 3.0 0.0 -1.0 0.0 0.0 0.0 -2.0 1.0 0.0 0.0 0.0 0.0 0.0 -6.0 5.0 0.0 -3.0 2.0 6.0 0.0 -4.0 3.0 0.0 -1.0 -5.0 4.0 0.0 -2.0 1.0 0.0

source

ModernRobotics.se3ToVec — Method.

se3ToVec(se3mat)

Converts an se3 matrix into a spatial velocity vector.

Examples

```jldoctest julia> se3ToVec([0 -3 2 4; 3 0 -1 5; -2 1 0 6; 0 0 0 0]) 1×6 Array{Int64,2}: 1 2 3 4 5 6

source

ModernRobotics.so3ToVec — Method.

so3ToVec(so3mat)

Converts an so(3) representation to a 3-vector.

Examples

```jldoctest julia> so3ToVec([0 -3 2; 3 0 -1; -2 1 0]) 3-element Array{Int64,1}: 1 2 3

source